requirements.txt
langchain
langchain-community
langchain-chroma
langchain-openai
chromadb
pypdf
sentence-transformers
torch
numpy
openai
tiktoken
pdfplumber
python-dotenv